feat: add delete exception and new DataBaseException

+++ src/main/java/com/calebecarozzi/javaordermanagementapi/services/UserService.java
@@ -2,8 +2,11 @@ package com.calebecarozzi.javaordermanagementapi.services;
 
 import com.calebecarozzi.javaordermanagementapi.entities.User;
 import com.calebecarozzi.javaordermanagementapi.repositories.UserRepository;
+import com.calebecarozzi.javaordermanagementapi.services.exceptions.DataBaseException;
 import com.calebecarozzi.javaordermanagementapi.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -31,7 +34,15 @@ public class UserService {
     }
 
     public void delete(Long id){
-        repository.deleteById(id);
+        try{
+            if (!repository.existsById(id)) {
+                throw new ResourceNotFoundException(id);
+            }
+            repository.deleteById(id);
+            repository.deleteById(id);
+        }catch(DataIntegrityViolationException e){
+            throw new DataBaseException(e.getMessage());
+        }
     }
 
     public User update(Long id, User obj){
